+++ doGRT/ViewControllers/UIViewController+GRTGtfsUpdater.m
@@ -104,7 +104,7 @@ static id theOneRequestUpdate = nil;
 		NSInteger date = endDate.integerValue;
 		labelBarItemSet = [ITLabelBarItemSet labelBarItemSetWithDismissTarget:self andAction:@selector(hideBarItemSet:)];
 		labelBarItemSet.textLabel.text = @"Update Succeed!";
-		labelBarItemSet.detailTextLabel.text = [NSString stringWithFormat:@"New schedule valid until %ld/%ld/%ld", (date / 100) % 100, date % 100, date / 10000];
+		labelBarItemSet.detailTextLabel.text = [NSString stringWithFormat:@"New schedule valid until %ld/%ld/%ld", (long)(date / 100) % 100, (long)date % 100, (long)date / 10000];
 	}
 	else {
 		labelBarItemSet = [ITConfirmationBarItemSet confirmationBarItemSetWithTarget:self andConfirmAction:@selector(startUpdate:) andDismissAction:@selector(hideBarItemSet:)];
